+++ test-engine/src/main/java/com/aws/jverify/testengine/JVerifyTestEngine.java
@@ -308,7 +308,8 @@ public class JVerifyTestEngine extends HierarchicalTestEngine<EngineExecutionCon
                         //"--wait-for-debugger",
                 },
                 annotation.verifyByDefault(),
-                annotation.continueOnErrors()
+                annotation.continueOnErrors(),
+                null
         );
     }
