+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/simplifications/VerifyAnnotationCompiler.java
@@ -1,7 +1,6 @@
 package com.aws.jverify.verifier.compiler.simplifications;
 
 import com.aws.jverify.Verify;
-import com.aws.jverify.verifier.PositionFilter;
 import com.aws.jverify.verifier.VerifierOptions;
 import com.aws.jverify.verifier.compiler.Reporter;
 import com.aws.jverify.verifier.compiler.dafnygenerator.base.BaseDafnyGenerator;
@@ -14,8 +13,6 @@ import com.sun.tools.javac.util.List;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * Depends on MethodOrLoopContractCompiler
@@ -24,7 +21,6 @@ public class VerifyAnnotationCompiler extends TreeScanner {
     private final JVerifyUtils jverifyUtils;
     private final VerifierOptions options;
     private final Reporter reporter;
-    public final Set<Symbol.MethodSymbol> removedImplementations = new HashSet<>();
 
     public VerifyAnnotationCompiler(Context context) {
         context.put(VerifyAnnotationCompiler.class, this);
@@ -72,7 +68,7 @@ public class VerifyAnnotationCompiler extends TreeScanner {
 
     @Override
     public void visitMethodDef(JCTree.JCMethodDecl tree) {
-        boolean shouldVerify = processVerifyAnnotationAndPop(tree.sym);
+        boolean shouldVerify = processVerifyAnnotationAndPop(tree, tree.sym);
         if (!shouldVerify) {
             removeImplementation(tree);
         }
@@ -80,14 +76,16 @@ public class VerifyAnnotationCompiler extends TreeScanner {
     }
 
     public void removeImplementation(JCTree.JCMethodDecl tree) {
+        if (tree.body == null || tree.body.getStatements().isEmpty()) {
+            return;
+        }
         var contractBlock = MethodOrLoopContractCompiler.getContractBlock(tree);
         tree.body.stats = List.of(contractBlock, jverifyUtils.contractThrow());
-        removedImplementations.add(tree.sym);
     }
 
-    public boolean processVerifyAnnotationAndPop(AnnoConstruct annoConstruct) {
+    public boolean processVerifyAnnotationAndPop(JCTree node, AnnoConstruct annoConstruct) {
         processVerifyAnnotation(annoConstruct);
-        boolean shouldVerify = shouldVerify();
+        boolean shouldVerify = shouldVerify() && applyPositionFilter(node);
         shouldVerifies.pop();
         return shouldVerify;
     }
@@ -104,7 +102,7 @@ public class VerifyAnnotationCompiler extends TreeScanner {
         throw new RuntimeException("shouldVerify should never be empty");
     }
     
-    private boolean shouldVerify2(JCTree node) {
+    private boolean applyPositionFilter(JCTree node) {
         var filter = options.positionFilter();
         if (filter == null) {
             return true;
